Create Linear_Reg
This model proved to be most practical with accuracy up to ~96%. It is logical that in real life too, Marketing and R&D Spend is linearly related to the profit of a company. Also this model is not strictly 'Linear Regression' per se, it is 'Multiple Linear Regression' as there are two input parameters.